Move messanger (#498)
