+++ src/messages/__init__.py
@@ -1,2 +1,3 @@
 from .generic_api import GenericApi
 from .discord_embeds import DiscordEmbeds
+from .messenger import Messenger
