+++ src/messages/messenger.py
@@ -2,8 +2,9 @@ from dataclasses import dataclass
 from config import Config
 import numpy as np
 
-from api.generic_api import GenericApi
-from api.discord_embeds import DiscordEmbeds
+from messages.generic_api import GenericApi
+from messages.discord_embeds import DiscordEmbeds
+
 
 class Messenger:
     def __init__(self):
@@ -17,13 +18,13 @@ class Messenger:
 
     def send_item(self, item: str, image:  np.ndarray, location: str):
         self._message_api.send_item(item, image, location)
-        
+
     def send_death(self, location: str, image_path: str = None):
         self._message_api.send_death(location, image_path)
-        
+
     def send_chicken(self, location: str, image_path: str = None):
         self._message_api.send_chicken(location, image_path)
-        
+
     def send_stash(self):
         self._message_api.send_stash()
